+++ changelog.md
@@ -5,6 +5,7 @@
 - allow exporting timeline in YouTube format
 - use dynamic sizing for modals (win32/macos)
 - app now supports viewing the video list, user can remove a video from the imported list
+- support customized intro / outro imports
 
 ### v1.6.4